+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_2_1)
+### Added
+- "Unity Fix" Modifier that fixes rotations missmatch between blender and unity
+
+### Changed
+- The default unity FBX preset now has the bake transforms experimental option disabled
+
 ## [2.2.0](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_2_0)
 ### Added
 - Option for merge meshes to merge UVs by index instead of by name